+++ changelog.md
@@ -2,6 +2,7 @@
 - Debug Render Pass buffers.
 - More WGSL function autocomplete constants and utils.
 - Fixed show shader code in mobile.
+- `iMouse.scroll` uniform.
 
 # 0.17.2
 - Updated LexGUI to `8.4.1`.